chore(csv): remove old CSV functions
